refactor(backend)!: Remove parameter indexes
This commit removes `IParameterShape#ParameterIndex`

The rationale behind this is that the parameter was only added in the first place to ensure that parameters were parsed in order, however this is moot as the order of the command parameters is determined by the order they're added/created in, anyway, and thus ordering them is just unneccesary overhead.

BREAKING-CHANGE: ParameterIndex was part of the public-facing API and has since been removed. Consumers should store an index when iterating over the command parameters, or use `.IndexOf(T)` on the collection.