fix: strings with excess characters are dropped by TimeSpanParser
TimeSpanParser currently ignores excess characters in front of or
after the matches. This means unexpected input like x60sx will lead to
a match of 60 s. The same, worse so, goes for - in front of the
TimeSpan.

This commit checks for excess characters, and returns a parsing error.
In addition it implements support for negative time spans. Negative
time spans are already supported in cases where TimeSpan.TryParse
parses the TimeSpan, but not with the custom format.